Move traits to sensible location
