--- core/editor/src/hint.rs
@@ -1,5 +1,0 @@
-use std::collections::HashMap;
-
-pub trait Hint {
-	fn hints(&self) -> HashMap<String, String>;
-}
